# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个超级订单管理系统的后端 API 文档",
        "title": "超级订单管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8173",
    "basePath": "/api",
    "paths": {
...
        "/api/order": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/order-item": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrderItem"
                ],
                "summary": "创建订单明细",
                "parameters": [
                    {
                        "description": "订单明细信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateOrderItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/order-item/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrderItem"
                ],
                "summary": "获取订单明细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单明细ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrderItem"
                ],
                "summary": "更新订单明细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单明细ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "订单明细信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateOrderItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrderItem"
                ],
                "summary": "删除订单明细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单明细ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/order/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "获取订单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "订单状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/order/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "获取订单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "更新订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "订单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "删除订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/order/{id}/items": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "获取订单明细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateOrderItemRequest": {
            "type": "object",
            "required": [
                "cost_price",
                "order_id",
                "product_name",
                "quantity",
                "sale_price",
                "sku_code",
                "total_cost_amount",
                "total_sale_amount"
            ],
            "properties": {
                "cost_price": {
                    "type": "number"
                },
                "order_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "sale_price": {
                    "type": "number"
                },
                "sku_code": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "integer"
                },
                "total_cost_amount": {
                    "type": "number"
                },
                "total_sale_amount": {
                    "type": "number"
                }
            }
        },
        "handler.CreateOrderRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "items",
                "order_date"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/handler.OrderItemRequest"
                    }
                },
                "order_date": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                }
            }
        },
...
        "handler.OrderItemRequest": {
            "type": "object",
            "required": [
                "cost_price",
                "product_name",
                "quantity",
                "sale_price",
                "sku_code",
                "total_cost_amount",
                "total_sale_amount"
            ],
            "properties": {
                "cost_price": {
                    "type": "number"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "sale_price": {
                    "type": "number"
                },
                "sku_code": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "integer"
                },
                "total_cost_amount": {
                    "type": "number"
                },
                "total_sale_amount": {
                    "type": "number"
                }
            }
        },
...
        "handler.UpdateOrderItemRequest": {
            "type": "object",
            "required": [
                "cost_price",
                "product_name",
                "quantity",
                "sale_price",
                "sku_code",
                "total_cost_amount",
                "total_sale_amount"
            ],
            "properties": {
                "cost_price": {
                    "type": "number"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "sale_price": {
                    "type": "number"
                },
                "settled_amount": {
                    "type": "number"
                },
                "sku_code": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "integer"
                },
                "total_cost_amount": {
                    "type": "number"
                },
                "total_sale_amount": {
                    "type": "number"
                }
            }
        },
        "handler.UpdateOrderRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "order_date"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "order_date": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_cost_amount": {
                    "type": "number"
                },
                "total_sale_amount": {
                    "type": "number"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}